--- a/api/src/App/Dependencies.php
+++ b/api/src/App/Dependencies.php
@@ -7,7 +7,9 @@ use Psr\Container\ContainerInterface;
 use App\Handler\ApiError;
 use App\Handler\RBAC;
 use App\Repository\DepartmentRepository;
+use App\Repository\EventRepository;
 use App\Service\DepartmentService;
+use App\Service\EventService;
 use Slim\App;
 
 /* setupAPIDependencies */
@@ -40,6 +42,12 @@ function setupAPIDependencies(App $app, array $settings)
     $container['DepartmentService'] = function ($c): DepartmentService {
         return new DepartmentService(
             new DepartmentRepository($c['db'])
+        );
+    };
+
+    $container['EventService'] = function ($c): EventService {
+        return new EventService(
+            new EventRepository($c['db'])
         );
     };