--- a/api/src/App/Dependencies.php
+++ b/api/src/App/Dependencies.php
@@ -8,8 +8,10 @@ use App\Handler\ApiError;
 use App\Handler\RBAC;
 use App\Repository\DepartmentRepository;
 use App\Repository\EventRepository;
+use App\Repository\MemberRepository;
 use App\Service\DepartmentService;
 use App\Service\EventService;
+use App\Service\MemberService;
 use Slim\App;
 
 /* setupAPIDependencies */
@@ -48,6 +50,12 @@ function setupAPIDependencies(App $app, array $settings)
     $container['EventService'] = function ($c): EventService {
         return new EventService(
             new EventRepository($c['db'])
+        );
+    };
+
+    $container['MemberService'] = function ($c): MemberService {
+        return new MemberService(
+            new MemberRepository($c['db'])
         );
     };