This PR includes performance improvement code that is built upon the Controller -> Service -> Repo pattern that was introduced in this PR: https://github.com/CON-In-A-Box/CIAB-Portal/pull/647 

This code more fully leverages the pattern and demonstrates how we can leverage it to make requests to the DB and cleanly format the response.

There are probably some additional optimizations that could be done where I am currently looping through departments and members by better use of the PHP arrays to leverage keys for the queries. After a few quick tests locally I'm not sure that it will matter too much, but we could consider it for the future.

There are no tests added for the specific service classes here, but I was able to verify the output matched between the original call and new call using JSON responses from the API and running through a JS test (not committed).

In cases where functions are added for convenience (such as `findDepartmentsById`), the full object was not pulled from the database. Instead, only enough for the API to perform the next step in its processing was retrieved and used.

I did some quick sample testing with ~33 staff as the highest in a single division and saw:

- Original Request was averaging about 335ms per request
- Updated Request was averaging about 94ms per request

I then added more staff manually to see how much of a performance boost we were getting with this as the size of staff per department increased. I tested again with 77 staff and saw these results:

- Original request for that division was 632ms for the first attempt, 580ms for the second.
- Updated request for that division was 85ms for the first attempt, 84ms for the second.

I added a few more so that I was over 100 staff in a single division and saw:

- Original request around 730ms or so
- Updated request was still 85ms