@rkaiser-dev 2023-11-29
> When doing API changes I feel like it is better try to to refer to the API changes instead of the feature request.

To make sure I'm on the same page and understanding, are "API changes" in this context meaning, the changes required to get the server's API output in parity with the existing implementation, or does it mean the programmatic APIs we are calling on the server to eventually get there?