@rkaiser-dev 2023-12-04
I rebased onto the latest changes from https://github.com/CON-In-A-Box/CIAB-Portal/pull/647 and adjusted this to use `buildResource` in its controller as well. Everything looked good locally and all of the changes from the other PR also looked good with these changes as well.

I will take a look at this and rebase immediately once https://github.com/CON-In-A-Box/CIAB-Portal/pull/647 has been merged into Main