@rkaiser-dev 2023-12-04
Everything has been moved accordingly. I also added some caching, although this code will not leverage that as it only calls for the current event and then uses that elsewhere.